+++ src/utils/useRestaurantMenu.js
@@ -1,0 +1,21 @@
+import { useState, useEffect } from "react";
+
+import { RESTAURANT_MENU_URL } from "./constants";
+
+const useRestaurantMenu = (resId) => {
+  const [resInfo, setResInfo] = useState(null);
+
+  useEffect(() => {
+    fetchMenu();
+  }, []);
+
+  const fetchMenu = async () => {
+    const data = await fetch(`${RESTAURANT_MENU_URL}${resId}`);
+    const jsonData = await data.json();
+    setResInfo(jsonData.data);
+  };
+
+  return resInfo;
+};
+
+export default useRestaurantMenu;
